+++ lib/FinancialConnections/Account.php
@@ -16,6 +16,7 @@ namespace Stripe\FinancialConnections;
  * @property string $category The type of the account. Account category is further divided in <code>subcategory</code>.
  * @property int $created Time at which the object was created. Measured in seconds since the Unix epoch.
  * @property null|string $display_name A human-readable name that has been assigned to this account, either by the account holder or by the institution.
+ * @property null|\Stripe\StripeObject $inferred_balances_refresh The state of the most recent attempt to refresh the account's inferred balance history.
  * @property string $institution_name The name of the institution that holds this account.
  * @property null|string $last4 The last 4 digits of the account number. If present, this will be 4 numeric characters.
  * @property bool $livemode Has the value <code>true</code> if the object exists in live mode or the value <code>false</code> if the object exists in test mode.
@@ -24,13 +25,16 @@ namespace Stripe\FinancialConnections;
  * @property null|string[] $permissions The list of permissions granted by this account.
  * @property string $status The status of the link to the account.
  * @property string $subcategory <p>If <code>category</code> is <code>cash</code>, one of:</p><p>- <code>checking</code> - <code>savings</code> - <code>other</code></p><p>If <code>category</code> is <code>credit</code>, one of:</p><p>- <code>mortgage</code> - <code>line_of_credit</code> - <code>credit_card</code> - <code>other</code></p><p>If <code>category</code> is <code>investment</code> or <code>other</code>, this will be <code>other</code>.</p>
+ * @property null|string[] $subscriptions The list of data refresh subscriptions requested on this account.
  * @property string[] $supported_payment_method_types The <a href="https://stripe.com/docs/api/payment_methods/object#payment_method_object-type">PaymentMethod type</a>(s) that can be created from this account.
+ * @property null|\Stripe\StripeObject $transaction_refresh The state of the most recent attempt to refresh the account transactions.
  */
 class Account extends \Stripe\ApiResource
 {
     const OBJECT_NAME = 'financial_connections.account';
 
     use \Stripe\ApiOperations\All;
+    use \Stripe\ApiOperations\NestedResource;
     use \Stripe\ApiOperations\Retrieve;
 
     const CATEGORY_CASH = 'cash';
@@ -100,5 +104,21 @@ class Account extends \Stripe\ApiResource
         $this->refreshFrom($response, $opts);
 
         return $this;
+    }
+
+    const PATH_INFERRED_BALANCES = '/inferred_balances';
+
+    /**
+     * @param string $id the ID of the account on which to retrieve the financial connections account inferred balances
+     * @param null|array $params
+     * @param null|array|string $opts
+     *
+     * @throws \Stripe\Exception\ApiErrorException if the request fails
+     *
+     * @return \Stripe\Collection<\Stripe\FinancialConnections\FinancialConnectionsAccountInferredBalance> the list of financial connections account inferred balances
+     */
+    public static function allInferredBalances($id, $params = null, $opts = null)
+    {
+        return self::_allNestedResources($id, static::PATH_INFERRED_BALANCES, $params, $opts);
     }
 }